Modify default behavior of maxhops to truncate
Since maxhops is currently undocumented, I made a small change to move
existing error behavior to strict mode, in line with other strict mode
errors.

In non-strict mode, maxhops will truncate the list down to whatever the
maxhops value is set to. This allows for handling of scenarios where the
set of CIDR ranges is either unknown, too long to list, or too
frequently changing. Instead you can set your CIDR range to 0.0.0.0/0
and, e.g. for a Caddy server behind a single load balancer, you can set
maxhops to 1. Any IPs to the left of the 1 IP address allowed will
simply be truncated. This allows you to trivially eliminate IPs internal
to other networks which are provided by forward proxies, and
consistently obtain the IP you're actually looking for in upstream code.

While it's a breaking change, it's a breaking change to an undocumented
feature and this seems like a better behavioral fit for a plugin named
'realip', and it doesn't prevent people from getting the original
behavior if they so desire.

Also documented the full behavior of maxhops and added a bunch more
tests, including explicit tests of strict mode.